// x-pack/platform/plugins/shared/actions/server/types.ts
export type ActionTypeConfig = Record<string, unknown>;
export type ActionTypeSecrets = Record<string, unknown>;
export type ActionTypeParams = Record<string, unknown>;
export type ConnectorTokenClientContract = PublicMethodsOf<ConnectorTokenClient>;

import type { Connector, ConnectorWithExtraFindData } from './application/connector/types';
import type { ActionExecutionSource, ActionExecutionSourceType } from './lib';
import type { ConnectorUsageCollector } from './usage';

export interface Services {
  savedObjectsClient: SavedObjectsClientContract;
  scopedClusterClient: ElasticsearchClient;
  connectorTokenClient: ConnectorTokenClient;
}

export interface UnsecuredServices {
  savedObjectsClient: ISavedObjectsRepository;
  scopedClusterClient: ElasticsearchClient;
  connectorTokenClient: ConnectorTokenClient;
}
// ...
export interface ActionTypeExecutorOptions<
  Config extends Record<string, unknown>,
  Secrets extends Record<string, unknown>,
  Params
> {
  actionId: string;
  services: Services | UnsecuredServices;
  config: Config;
  secrets: Secrets;
  params: Params;
  logger: Logger;
  globalAuthHeaders?: Record<string, AxiosHeaderValue>;
  taskInfo?: TaskInfo;
  configurationUtilities: ActionsConfigurationUtilities;
  source?: ActionExecutionSource<unknown>;
  request?: KibanaRequest;
  connectorUsageCollector: ConnectorUsageCollector;
  connectorTokenClient?: ConnectorTokenClientContract;
  signal?: AbortSignal;
  authMode?: AuthMode;
  profileUid?: string;
}
// ...
export type ExecutorType<
  Config extends Record<string, unknown>,
  Secrets extends Record<string, unknown>,
  Params,
  ResultData
> = (
  options: ActionTypeExecutorOptions<Config, Secrets, Params>
) => Promise<ActionTypeExecutorResult<ResultData>>;

type Validator<T> = Pick<z3.ZodType, 'parse'> | Pick<z4.ZodType, 'parse'>;
export interface ValidatorType<T> {
  schema: Validator<T>;
  customValidator?: (value: T, validatorServices: ValidatorServices) => void;
}

export interface ValidatorServices {
  configurationUtilities: ActionsConfigurationUtilities;
}
// ...
export type ActionType<
  Config extends ActionTypeConfig = ActionTypeConfig,
  Secrets extends ActionTypeSecrets = ActionTypeSecrets,
  Params extends ActionTypeParams = ActionTypeParams,
  ExecutorResultData = void
> =
  | ClassicActionType<Config, Secrets, Params, ExecutorResultData>
  | WorkflowActionType<Config, Secrets, Params, ExecutorResultData>;

export interface ActionTypeCoreFields<
  Config extends ActionTypeConfig = ActionTypeConfig,
  Secrets extends ActionTypeSecrets = ActionTypeSecrets,
  Params extends ActionTypeParams = ActionTypeParams
> {
  id: string;
  name: string;
  maxAttempts?: number;
  minimumLicenseRequired: LicenseType;
  supportedFeatureIds: string[];
  isSystemActionType?: boolean;
  source?: ActionTypeSource;
  subFeature?: SubFeature;
  isDeprecated?: boolean;
  /**
   * Allows multiple instances of the same system action in a single rule.
   * By default, system actions can only be used once per rule.
   * Set to true to allow the same system action connector to be used multiple times.
   * Only applies to system actions (isSystemActionType: true).
   */
  allowMultipleSystemActions?: boolean;
  /**
   * Additional Kibana privileges to be checked by the actions framework.
   * Use it if you want to perform extra authorization checks based on a Kibana feature.
   * For example, you can define the privileges a users needs to have to execute
   * a Case or OsQuery system action.
   *
   * The list of the privileges follows the Kibana privileges format usually generated with `security.authz.actions.*.get(...)`.
   *
   * It only works with system actions and only when executing an action.
   * For all other scenarios they will be ignored
   */
  getKibanaPrivileges?: (args?: {
    params?: Params;
    source?: ActionExecutionSourceType;
  }) => string[];
  // Headers that should be added to every Axios request made by this action type
  globalAuthHeaders?: Record<string, AxiosHeaderValue>;
  renderParameterTemplates?: RenderParameterTemplates<Params>;
  getService?: (params: ServiceParams<Config, Secrets>) => SubActionConnector<Config, Secrets>;
  preSaveHook?: (params: PreSaveConnectorHookParams<Config, Secrets>) => Promise<void>;
  postSaveHook?: (params: PostSaveConnectorHookParams<Config, Secrets>) => Promise<void>;
  postDeleteHook?: (params: PostDeleteConnectorHookParams<Config, Secrets>) => Promise<void>;
}

export type WorkflowActionType<
  Config extends ActionTypeConfig = ActionTypeConfig,
  Secrets extends ActionTypeSecrets = ActionTypeSecrets,
  Params extends ActionTypeParams = ActionTypeParams,
  ExecutorResultData = void
> = ActionTypeCoreFields<Config, Secrets, Params> & {
  executor?: ExecutorType<Config, Secrets, Params, ExecutorResultData>;
  validate: {
    params?: ValidatorType<Params>;
    config: ValidatorType<Config>;
    secrets: ValidatorType<Secrets>;
    connector?: (config: Config, secrets: Secrets) => string | null;
  };
};

export type ClassicActionType<
  Config extends ActionTypeConfig = ActionTypeConfig,
  Secrets extends ActionTypeSecrets = ActionTypeSecrets,
  Params extends ActionTypeParams = ActionTypeParams,
  ExecutorResultData = void
> = ActionTypeCoreFields<Config, Secrets, Params> & {
  executor: ExecutorType<Config, Secrets, Params, ExecutorResultData>;
  validate: {
    params: ValidatorType<Params>;
    config: ValidatorType<Config>;
    secrets: ValidatorType<Secrets>;
    connector?: (config: Config, secrets: Secrets) => string | null;
  };
};